fix(/teacher/studentinfo;/description): fix the bug of teacher pages and add Description page
