+++ src/pages/Teacher/StudentInfo/index.module.less
@@ -6,4 +6,14 @@
       background: #00060b !important;
     }
   }
+
+  .info {
+    background-color: #4a7658;
+    color: #fff;
+    font-size: 1rem;
+    display: flex;
+    justify-content: space-evenly;
+    line-height: 4rem;
+    letter-spacing: 1px;
+  }
 }
